GFS2: Don't brelse rgrp buffer_heads every allocation
[ Upstream commit 39b0f1e9290880a6c905f639e7db6b646e302a4f ]

This patch allows the block allocation code to retain the buffers
for the resource groups so they don't need to be re-read from buffer
cache with every request. This is a performance improvement that's
especially noticeable when resource groups are very large. For
example, with 2GB resource groups and 4K blocks, there can be 33
blocks for every resource group. This patch allows those 33 buffers
to be kept around and not read in and thrown away with every
operation. The buffers are released when the resource group is
either synced or invalidated.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Reviewed-by: Steven Whitehouse <swhiteho@redhat.com>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>
[Lee: Back-ported to solve a build dependency]
Signed-off-by: Lee Jones <lee.jones@linaro.org>
Change-Id: I67323c027fce5034afe0a3dcf44ea7524aa468a9